==> 4_Dashboard.ipynb <==
from myfunctions import *
import ipywidgets as wg
from IPython.display import display

# import edges data needed for the Graph and transform into a list
from hdfs3 import HDFileSystem
hdfs = HDFileSystem(host='hdfs://iccluster044.iccluster.epfl.ch', port=8020, user='ebouille')
files = hdfs.glob('/user/eckes/edges.parquet/*.parquet')
df = pd.DataFrame()
for file in files:
    with hdfs.open(file) as f:
        df = df.append(pd.read_parquet(f))
edges_list = [tuple(r) for r in df.to_numpy()]

# import walking table needed for the Graph 
from hdfs3 import HDFileSystem
hdfs = HDFileSystem(host='hdfs://iccluster044.iccluster.epfl.ch', port=8020, user='ebouille')
files = hdfs.glob('/user/eckes/walking_table.parquet/*.parquet')
walking_table = pd.DataFrame()
for file in files:
    with hdfs.open(file) as f:
        walking_table = walking_table.append(pd.read_parquet(f))

# define function we will use in the widget
def plot_itinerary(start_node, goal_node, end_time, confidence_proba):
    graph=Graph(start_node, goal_node, edges_list, end_time, confidence_proba, walking_table)
    df_itinerary=graph.clean_itinary()
    return df_itinerary

# define variables that need input
Departure=wg.Text(value='Departure')
Destination=wg.Text(value='Destination')
Arrival_time=wg.Text(value='Arrival time')
prob=wg.Dropdown(description='Probability %:', options=[80, 85, 90, 95])

# create the widget
wg.interact_manual(plot_itinerary,
                   start_node = Departure,
                   goal_node = Destination,
                   end_time = Arrival_time,
                   confidence_proba = prob)
